backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
Pillow==11.0.0
python-decouple==3.8
gunicorn==23.0.0
django-storages==1.14.4
boto3==1.35.53
